refactor(aethe-cli): Remove dependency on aether-evals and split out aether-schemas pkg
