+++ packages/aether-evals/src/agents/mod.rs
@@ -3,7 +3,6 @@ mod docker;
 mod docker_agent;
 mod env;
 mod fake_agent;
-mod image_builder;
 mod transcript;
 
 pub use agent::{Agent, AgentConfig, RunError};
@@ -13,7 +12,6 @@ pub use docker_agent::{
 };
 pub use env::{CONTAINER_AETHER_HOME, default_eval_env_vars};
 pub use fake_agent::FakeAgent;
-pub use image_builder::{ImageBuildError, ImageBuildRequest, build_images};
 
 pub(crate) use agent::build_task_prompt;
-pub(crate) use transcript::{TRANSCRIPT_PAYLOAD_CHARS, get_transcript_line, is_terminal, truncate_chars};
+pub(crate) use transcript::{TRANSCRIPT_PAYLOAD_CHARS, get_transcript_line, is_terminal};
