--- packages/internal-evals/examples/edit-notes.eval.json
@@ -1,60 +1,0 @@
-{
-  "name": "edits_notes_file",
-  "docker": {
-    "file": "Dockerfile",
-    "context": "../../..",
-    "image": "aether-sandbox:latest"
-  },
-  "agent": {
-    "command": ["/usr/local/bin/aether-eval-agent"]
-  },
-  "task": {
-    "prompt": "Use the coding MCP tools to update notes.txt. Read the file first, then call coding__edit_file exactly once to replace only the first 'alpha' with 'beta'. Do not replace the second 'alpha'.",
-    "workspace": {
-      "files": {
-        "notes.txt": "alpha\nalpha\n"
-      }
-    }
-  },
-  "expect": {
-    "toolCalls": {
-      "coding__read_file": {
-        "atLeast": 1
-      },
-      "coding__edit_file": {
-        "exactly": 1
-      }
-    },
-    "files": {
-      "notes.txt": "beta\nalpha\n"
-    },
-    "judge": {
-      "model": "zai:glm-5.1",
-      "instructions": "Grade whether this would be accepted by a maintainer.",
-      "contextFiles": ["notes.txt"],
-      "criteria": [
-        {
-          "id": "behavior",
-          "description": "The agent replaces only the first alpha with beta and leaves the second alpha unchanged.",
-          "blocking": true,
-          "weight": 3.0,
-          "threshold": 1.0
-        },
-        {
-          "id": "scope",
-          "description": "The agent avoids unrelated file changes and extra refactors.",
-          "blocking": true,
-          "weight": 1.0,
-          "threshold": 0.8
-        },
-        {
-          "id": "clarity",
-          "description": "The final response clearly explains the completed change.",
-          "blocking": false,
-          "weight": 0.5,
-          "threshold": 0.7
-        }
-      ]
-    }
-  }
-}
